requirements.txt
appnope==0.1.0
decorator==4.0.11
Django==1.11.3
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.6.3
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
mongoengine==0.13.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.2
Pygments==2.2.0
pymongo==3.4.0
pytz==2017.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
wheel==0.24.0
